+++ src/main/java/lotto/controller/LottoController.java
@@ -5,12 +5,12 @@ import lotto.model.domain.CompareResult;
 import lotto.Lotto;
 import lotto.model.dto.ResultForView;
 import lotto.model.service.LottoComparator;
-import lotto.model.service.LottoPublisher;
+import lotto.model.service.lottopublisher.LottoPublisher;
 import lotto.model.service.PrizeCalculator;
 import lotto.model.domain.PublishedLotto;
 import lotto.model.domain.PurchasePrice;
 import lotto.model.domain.WinningLotto;
-import lotto.model.converter.LottoConverter;
+import lotto.model.service.LottoConverter;
 import lotto.model.dto.PublishedLottoDto;
 import lotto.util.InputUtil;
 import lotto.model.validator.LottoNumberValidator;
