+++ src/main/java/lotto/model/validator/LottoNumberValidator.java
@@ -1,6 +1,6 @@
 package lotto.model.validator;
 
-import static lotto.exception.ErrorMessage.INVALID_LOTTO_NUMBER_RANGE;
+import static lotto.model.constant.ErrorMessage.INVALID_LOTTO_NUMBER_RANGE;
 import static lotto.model.constant.LottoConstant.MAX_NUMBER;
 import static lotto.model.constant.LottoConstant.MIN_NUMBER;
